requirement.txt
click==6.7
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
SQLAlchemy==1.2.11
Werkzeug==0.14.1
WTForms==2.2.1
pymysql
cos-python-sdk-v5
qcloud-python-sts
